src/dspygen/modules/chat_bot_cli.py

import typer
app = typer.Typer()

@app.command(name="chat")
def chat():
    """Start a chat session with the ChatBot."""
    # Command logic goes here
    print("This is the chat command.")

@app.command(name="clear")
def clear():
    """Clear the chat history."""
    # Command logic goes here
    print("This is the clear command.")

@app.command(name="history")
def history():
    """View the chat history."""
    # Command logic goes here
    print("This is the history command.")

@app.command(name="list")
def list():
    """List all available commands."""
    # Command logic goes here
    print("This is the list command.")

@app.command(name="settings")
def settings():
    """View or change ChatBot settings."""
    # Command logic goes here
    print("This is the settings command.")

@app.command(name="train")
def train():
    """Train the ChatBot with new data."""
    # Command logic goes here
    print("This is the train command.")

@app.command(name="version")
def version():
    """View the current ChatBot version."""
    # Command logic goes here
    print("This is the version command.")


if __name__ == "__main__":
    app()

