src/dspygen/modules/test_chat_bot_cli.py

import pytest
from typer.testing import CliRunner

from dspygen.modules.chat_bot_cli import app

runner = CliRunner()

def test_chat():
    result = runner.invoke(app, ["chat"])
    assert result.exit_code == 0
    assert "This is the chat command." in result.output  # Replace with specific expected output

def test_clear():
    result = runner.invoke(app, ["clear"])
    assert result.exit_code == 0
    assert "This is the clear command." in result.output  # Replace with specific expected output

def test_history():
    result = runner.invoke(app, ["history"])
    assert result.exit_code == 0
    assert "This is the history command." in result.output  # Replace with specific expected output

def test_list():
    result = runner.invoke(app, ["list"])
    assert result.exit_code == 0
    assert "This is the list command." in result.output  # Replace with specific expected output

def test_settings():
    result = runner.invoke(app, ["settings"])
    assert result.exit_code == 0
    assert "This is the settings command." in result.output  # Replace with specific expected output

def test_train():
    result = runner.invoke(app, ["train"])
    assert result.exit_code == 0
    assert "This is the train command." in result.output  # Replace with specific expected output

def test_version():
    result = runner.invoke(app, ["version"])
    assert result.exit_code == 0
    assert "This is the version command." in result.output  # Replace with specific expected output

